But not everyone agrees that putting JS in the footer is the best thing since Chuck Norris. After all, if your page is heavily dependent on JS, your user might see a blank page for a second before your JavaScript has the chance to execute and put cool stuff there, like a photo of Chuck Norris.
So, there might be some performance differences between putting JavaScript in the header versus the footer. But, our code should work equally well in either place, right? If I move the block javascripts up into my header, this should probably still work?

We still have 3 script tags, in the same order, just in a different spot.

Well... let's find out! Refresh! Then click delete. Ah we broke it! What happened?!

Running JavaScript Before the DOM

This may or may not be obvious to you, but it's worth mentioning: our browser executes JavaScript as soon as it sees it... which might be before some or all of the page has actually loaded. Our code is looking for all elements with the js-delete-rep-log class. Well, at this point, none of the HTML body has loaded yet, so it finds exactly zero elements.