Of course, the only problem is if someone comes along and decides:
Hey, you know what? We should really put our JavaScript in the footer! Chuck Norris told me it's better for performance.

Now, we have a different problem. In the source, jQuery once again lives at the absolute bottom. But when we refresh the page, error! Our browser immediately tells us that {dollar_sign} is not defined.

This comes from our code, which still lives in the middle of the page. And yea, it makes sense: as our browser loads the page, it sees the $, but has not yet downloaded jQuery: that script tag lives further down.

So there are two things we need to worry about. First, any JavaScript that I depend on needs to be included on the page before me. And actually, this will stop being true when we talk about module loaders in a future tutorial.

Second, before I try to select any elements with jQuery, I better make sure the DOM has loaded, which we can always guarantee with a {dollar_sign}(document).ready() block.