I'm feeling so good about our first click listener, let's add another! When I click anywhere on a row, I also want to log a message.
Back in the template, give the entire table a js class so we can select it. How about js-rep-log-table.

Down below, find that and look inside for the tbody tr elements. Then, .on('click') add a function that prints some fascinating text: console.log('row clicked').

Beautiful! Refresh and click the row. No surprises: we see 'row clicked'. But check this out: click the delete link. Hot diggity - two log messages! Of course it would do this! I clicked the delete link, but the delete link is inside of the row. Both things got clicked!