Cool! Let's play with this! First, let's clean up our code a bit and make a minor performance improvement. Add var {dollar_sign}table = {dollar_sign}('.js-rep-log-table'). Then below, instead of searching the entire page for these delete links, use .find() to only look inside that table.
Do the same below: {dollar_sign}table.find() and look for the tbody tr elements in that.

If you refresh now, it still works great. But some of you might be wondering about my variable name: {dollar_sign}table? For PHP developers, that looks weird... because, ya know, $ means something important in PHP. But in JavaScript, {dollar_sign} is not a special character. In fact, it's so not special that - if you want - you can even start a variable name with it. Madness! So the {dollar_sign} in {dollar_sign}table isn't doing anything special, but it is a fairly common convention to denote a variable that is a jQuery object.

It's nice because when I see {dollar_sign}table, I think:

Oh! This starts with a {dollar_sign}! Good show! I bet it's a jQuery object, and I can call find() or any other fancy jQuery method on it. Jolly good!
Now that we understand event bubbling, let's mess with it! Yes, we can actually stop the bubbling process... which is probably not something you want to do... but you might already be doing it accidentally.