So what is this e.target thing exactly? I mean, is it a string? Or an object? And what else can we do with it? Let's go digging! Add console.log(e.target).
And then, refresh! Ok, click on some delete links. Huh... it just prints out the HTML itself. So, it's a string?

Actually, no... our browser is kinda lying to us: e.target is a DOM Element object. Google for that and find the W3Schools(http://www.w3schools.com/jsref/dom_obj_all.asp) page all about it. You see, every element on the page is represented by a JavaScript object, a DOM Element object. My debugger is printing it like a string, but that's just for convenience... or inconvenience in this case. Nope, it's actually an object, with properties and methods that we can call. The W3Schools page shows all of this.