Ultimately all discused before means that we can say, {dollar_sign}(this).addClass('text-danger').That will always add the text-danger link to the anchor tag. That will always add the text-danger link to the anchor tag.
And finally, we can easily change our icon to a spinner! Just use {dollar_sign}(this).find('.fa') to find the icon inside of the anchor. Then, .removeClass('fa-trash'), .addClass('fa-spinner') and .addClass('fa-spin').

Refresh! Show me a spinner! There it is! It doesn't matter if we click the 'Delete' text or the trash icon itself.

So, use the this variable, it's your friend. But realize what's going on: this is just a shortcut to e.currentTarget. That fact is going to become critically important in just a little while.