Hmm, so after ajax call finishes, we probably want the entire row to disappear. Above the AJAX call, find the row with {dollar_sign}row = {dollar_sign}(this).closest('tr').
In other words, start with the link, and go up the DOM tree until you find a tr element. Oh, and reminder, this is {dollar_sign}row because this is a jQuery object! Inside success, say {dollar_sign}row.fadeOut() for just a little bit of fancy.

Ok, try that out! Refresh, delete my coffee cup and life is good. And if I refresh, it's truly gone. Oh, but dang, if I delete my cup of coffee record, the total weight at the bottom does not change. I need to refresh the page to do that. LAME! I'll re-add my coffee cup. Now, let's fix that!