Anyways, let's give our object a new method: an initialize key set to a function(). We'll call this when the page loads, and its job will be to attach all the event handlers for all the events that we need on our table. Give it a {dollar_sign}wrapper argument.
Before we do anything else, set that {dollar_sign}wrapper argument onto a property: this.{dollar_sign}wrapper = {dollar_sign}wrapper.
Yep, we just dynamically added a new property. This is the second time we've seen the this variable in JavaScript. And this time, it's more familiar: it refers to this object.

Next, copy our first listener registration code, but change {dollar_sign}table to this.{dollar_sign}wrapper. And instead of using a big ugly anonymous function, let's make this event call a new method on our object: this.handleRepLogDelete.

We'll add that in a moment.

Repeat this for the other event listener: copy the registration line, change {dollar_sign}table to this.{dollar_sign}wrapper, and then on click, call this.handleRowClick.

I already like it!