After initialize, create these methods! Add a key called, handleRepLogDelete set to a new function. Then go copy all of our original handler code, delete it, and put it in handleRepLogDelete body function. Make sure you have the, e argument exactly like before.
Do the same thing for our other method: handleRowClick set to a function() {}
I'm not using the, e argument, so I don't need to add it. Copy the console.log() line, delete it, and put it in handleRowClick body function.

There's one teenie detail I want you to notice: when we specify the event callback, this.handleRepLogDelete - we're not executing it.

I mean, there are no () on the end of it. Nope, we're simply passing the function as a reference to the on() function. If you forget and add (), things will get crazy.