Back in the (document).ready(), our job is really simple: find the {dollar_sign}table and then pass it to RepLogApp.initialize().
The cool thing about this approach is that now we have an entire object who's job is to work inside of this.{dollar_sign}wrapper.

Ok, let's try this! Go back and refresh! Hit delete! Ah, it fails!

Variable  is not defined.
The problem is inside of handleRepLogDelete. Ah, cool, this makes total sense. Before, we had a {dollar_sign}table variable defined above the function.