We just found out that, somehow, this.{dollar_sign}wrapper is not our jQuery object, it's undefined! 117 lines  app/Resources/views/lift/index.html.twig ... lines 1 - 64 {% block javascripts %} ... lines 66 - 67   <script>     var RepLogApp = { ... lines 70 - 82       handleRepLogDelete: function(e) { ... lines 84 - 93         var  = this.{dollar_sign}wrapper.find('.js-total-weight'); ... lines 95 - 103       }, ... lines 105 - 108     }; ... lines 110 - 114     </script> {% endblock %}
Rude! How is that even possible! The answer! Because JavaScript is weird, especially when it comes to the crazy this variable!

When this is not this

Here's the deal: whenever you are in a callback function, like the success callback of an AJAX call, the callback of an event listener, or even when passing a callback to the setTimeout() function, the this variable in your callback changes to be something else. And we already knew that! We know that this in our event handler is actually a reference to the DOM Element object that was clicked. So the this variable in handleRepLogDelete is not our RepLogApp object, even though we're inside of that object. Creepy!

We're going to talk a lot more about this situation... in a moment.