Since we're running out of items, let's add a few more!
Now that we have a fancy object, we can use it to get even more organized, by breaking big functions into smaller ones.

For example, we could create a new function called, updateTotalWeightLifted.

Instead of figuring out the total weight lifted in line 91 (index.php file) and doing the update down in the success callback we'll just call this method and have it do all that heavy lifting.

Add var totalWeight = 0.

Then I'll say, this.{dollar_sign}wrapper, which I can do because we're not in a callback function: this is our object. Then, .find to look for all tbody tr elements, and .each() to loop over them.

But stop! Notice that when you use .each(), you pass it a callback function! So guess what? Inside, this is no longer our RepLogApp object, it's something different. In this case, this is the individual tr DOM Element object that we're looping over in this moment.

Inside, add up all the total weights with totalWeight += {dollar_sign}(this).data() and read the data-weight attribute.

Finally use this.{dollar_sign}wrapper.find() to look for our js-total-weight element and set its HTML to totalWeight