How do I Know what this Is?
Here's the deal: when you call a function in JavaScript, you can choose to change what this is inside of that function when you call it. That means you could have one function and 10 different people could call your function and decide to set this to 10 different things.

Now, in reality, it's not that bad. But we do need to remember one rule of thumb: whenever you have a callback function - meaning someone else is calling a function after something happens - this will have changed. We've already seen this a lot: in the click functions, inside of .each(), inside of success and even inside of {dollar_sign}row.fadeOut():

 126 lines  app/Resources/views/lift/index.html.twig
... lines 1 - 64
{% block javascripts %}
... lines 66 - 67
  <script>
    var RepLogApp = {
      initialize: function({dollar_sign}wrapper) {
... lines 71 - 72
        this.{dollar_sign}wrapper.find('.js-delete-rep-log').on(
... line 74
          this.handleRepLogDelete
        );
        this.{dollar_sign}wrapper.find('tbody tr').on(
... line 78
           this.handleRowClick
        );
      },
      updateTotalWeightLifted: function() {
... line 84
        this.{dollar_sign}wrapper.find('tbody tr').each(function() {
          totalWeight += {dollar_sign}(this).data('weight');
        });
... lines 88 - 89
      },
      handleRepLogDelete: function(e) {
... lines 93 - 102
        {dollar_sign}.ajax({
... lines 104 - 105
          success: function() {
            {dollar_sign}row.fadeOut('normal', function() {
... line 108
              RepLogApp.updateTotalWeightLifted();
             });
           }
        });
... lines 113 - 117
      };
... lines 119 - 123
  </script>
{% endblock %}