But now, let's get tricky! Create a new variable called newThis and set it to an object with important stuff like cat set to meow and dog set to woof. To force newThis to be this inside our function, call the function indirectly with this.whatIsThis.call() and pass it newThis and the greeting, hello.
To force newThis to be this inside our function, call the function indirectly with this.whatIsThis.call() and pass it newThis and the greeting, hello.

Oh, and quick note: this.whatIsThis is, obviously, a function. But in JavaScript, functions are actually objects themselves! And there are a number of different methods that you can call on them, like .call(). The first argument to call() is the variable that should be used for this, followed by any arguments that should be passed to the function itself.

Refresh now and check this out! this is now our thoughtful cat, meow, dog, woof object. That is what is happening behind the scenes with your callback functions.

Now that we understand the magic behind this, how can we fix it? How can we guarantee that this is always our RepLogApp object when we're inside of it?