So how could we make this truly private? Well, you can't make methods or properties in an object private. BUT, you can make variables private, by taking advantage of variable scope. What I mean is, if I have access to the RepLogApp object, then I can call any methods on it. But if I didn't have access to this, or some other object, then of course I wouldn't be able to call any methods on it. I know that sounds weird, so let's do it!
At the bottom of this file, create another object called: var Helper = {}.

Commonly, we'll organize our code so that each file has just one object, like in PHP. But eventually, this variable won't be public - it's just a helper meant to be used only inside of this file.

I'll even add some documentation: this is private, not meant to be called from outside!

Just like before, give this an initialize, function with a  argument. And then say: this.{dollar_sign}wrapper = {dollar_sign}wrapper.

Move the calculateTotalWeight() function into this object, but take off the underscore.

Technically, if you have access to the Helper variable, then you're allowed to call calculateTotalWeight. Again, that whole _ thing is just a convention.