Now, go back into RepLogApp, find Helper, and remove the var.
You've probably been taught to never do this. And that's right! But you may not realize exactly what happens if you do.

Refresh again and open the window variable. Check this out! It's a little hard to find, but all of a sudden, there is a global Helper variable! So if you forget to say var - which you shouldn't - it makes that variable a global object, which means it's set on window.