How do we do that? Instead of setting Helper to {}, set it to a function. Let's set Helper to what was our initialize() method. Huh. So now, Helper is a function... But remember that functions are objects, so it's totally valid to add properties or methods to it.
Why would set our object to a function? Because now we are allowed to say this.helper = new Helper({dollar_sign}wrapper).

JavaScript does have the new keyword just like PHP! And you can use it once Helper is actually a function. This returns a new instance of Helper, which we set on a property.

In PHP, when you say new Helper(), PHP calls the constructor on your object, if you have one. The same happens here, the function is the constructor. At this point, we could create multiple Helper instances, each with their own {dollar_sign}wrapper.

Now, instead of using Helper in a static kind of way, we use its instance: this.helper.
Before we keep celebrating, let's try this. Go back, refresh, and delete one of our items! Huh, it worked... but the total didn't update. And, we have an error:

Uncaught TypeError: this.helper.calculateTotalWeight is not a function
That's odd! Why does it think our Helper doesn't have that key? The answer is all about the prototype.