In RepLogApp, when we try to call this.helper.calculateTotalWeight, for some reason, it doesn't think this is a function!
 71 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}) {
  window.RepLogApp = {
... lines 5 - 17
    updateTotalWeightLifted: function () {
      this..find('.js-total-weight').html(
        this.helper.calculateTotalWeight()
    );
  },
... lines 23 - 52
  };
... lines 54 - 69
})(window, jQuery);
But down below, we can plainly see: calculateTotalWeight is a function! What the heck is going on?

To find out, in initialize, let's log a few things: console.log(this.helper) and then Object.keys(this.helper)

The Object.keys method is an easy way to print the properties and methods inside an object.

Do the same thing for Helper and Object.keys(Helper).

Let's look at what the difference is between our instance of the Helper object and the Helper object itself.

Ok, find your browser, refresh, and check this out! There's the helper instance object, but check out the methods and properties on it: it has {dollar_sign}wrapper. Wait, so when we create a new Helper(), that instance object does have the {dollar_sign}wrapper property... but somehow it does not have a calculateTotalWeight method!

That's why we're getting the error. The question is why? Below, where we printed the upper-case 'H' Helper object, it prints out as a function, but in its keys, it does have one called calculateTotalWeight. Oooh, mystery!

This can be very confusing. So follow this next part closely and all the way to the end.

At this point, the calculateTotalWeight function is effectively still static. The only way that we can call that method is by saying Helper.calculateTotalWeight - by calling the method on the original, static object. We cannot call this method on the instantiated instance: we can't say this.helper.calculateTotalWeight(). It just doesn't work!